change target to long for b2b

+++ src/MLEC/models/SpanEmoB2B.py
@@ -54,10 +54,8 @@ class SpanEmoB2B(MLECDecoder):
         :return: loss, num_rows, y_pred, targets
         """
         inputs, targets, lengths, label_idxs = batch
-        inputs, num_rows = inputs.to(device).to(torch.int64), inputs.size(0)
-        label_idxs, targets = label_idxs[0].long().to(device), targets.float().to(
-            device
-        )
+        inputs, num_rows = inputs.long().to(device), inputs.size(0)
+        label_idxs, targets = label_idxs[0].long().to(device), targets.long().to(device)
         outputs = self.model(
             inputs,
             decoder_input_ids=targets,
